# comchill-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ComChill Core API Documentation",
        "description": "REST API endpoints for ComChill student messaging platform",
        "contact": {
            "email": "franciscohermas@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "ComChill Local Development API Server"
        }
    ],
    "paths": {
        "/api/ai/chat": {
            "post": {
                "tags": [
                    "AI"
                ],
                "summary": "Chat with the AI assistant",
                "description": "Processes a student message through the AI pipeline and returns the generated reply.",
                "operationId": "1448d990436dff4229f6fee0b52a9263",
                "responses": {
                    "200": {
                        "description": "AI response generated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/conversations/{conversation}/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List messages in a conversation",
                "description": "Returns a paginated list of messages for the specified conversation.",
                "operationId": "de51fd4a42a2b4a6706e448fdfae28a1",
                "responses": {
                    "200": {
                        "description": "Messages retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized access to this conversation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send a new message",
                "description": "Stores and sends a new message to the specified conversation.",
                "operationId": "25902563de55122ee7288085a684b2ab",
                "responses": {
                    "201": {
                        "description": "Message sent successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized access to this conversation"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/conversations/{conversation}/messages/read": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Mark conversation messages as read",
                "description": "Marks all messages in the specified conversation as read.",
                "operationId": "f28e2722df55cef6673a2ff190a80c4e",
                "responses": {
                    "200": {
                        "description": "Messages marked as read"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token to access protected routes",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}